// src/helpers/interfaces.ts
export interface Geo {
  lat: string;
  lng: string;
}
export interface Address {
  city: string;
  geo: Geo;
  street: string;
  suite: string;
  zipcode: string;
}

export interface Company {
  bs: string;
  catchPhrase: string;
  name: string;
}

export default interface User {
  id: number;
  name: string;
  username: string;
  email: string;
  phone: string;
  website: string;
  address: Address;
  company: Company;
}

export interface FetchState {
  data: User[];
  loading: boolean;
  error: string | null;
}